Refaktor. Flytt type til typefil
