+++ src/frontend/typer/barn.ts
@@ -7,6 +7,16 @@ export interface Barn {
     skalHaBarnepass: boolean;
 }
 
+export interface BarnMedAllInfo {
+    id: string;
+    alder: number;
+    fødselsdato: string;
+    fødselsnummer: string;
+    navn: string;
+    skalHaBarnepass: boolean;
+    passType?: PassType;
+}
+
 export enum PassType {
     BARNEHAGE_SFO_AKS = 'BARNEHAGE_SFO_AKS',
     ANDRE = 'ANDRE',
